requirements.txt
# requirements.txt (PİPER TTS İLE NİHAİ VE %100 ÇALIŞIR HALİ)

# ---------- Core ----------
python-dotenv==1.0.0
pydantic>=2.10,<3.0
pydantic-settings>=2.1,<3.0

# ---------- LLM (Ollama) ----------
aiohttp>=3.9,<4.0
ollama-python==0.1.2

# ---------- Speech-to-Text (HIZ ODAKLI) ----------
faster-whisper[ctranslate2]>=1.0.7,<2.0

# ---------- Text-to-Speech (YENİ, UYUMLU VE HIZLI) ----------
pyttsx3==2.90
pyobjc>=11.0 
# ---------- Translation (LOKAL) ----------
argostranslate>=1.9.0,<2.0

# ---------- Audio / VAD ----------
pyaudio==0.2.14
soundfile==0.12.1
webrtcvad==2.0.10
scipy>=1.15,<2.0

# ---------- CLI & Utils ----------
rich==13.7.0
loguru==0.7.2
tenacity==8.2.3